UGENE-7375. Check user forward and reverse primers. Namespace ReportUtils moved to struct PCRPrimerDesignTaskReportUtils in separate files.
